[FEATURE] Use a separate ParserTerm-object for parser & lazy-loading for relations in normal Term-object
Without lazy-loading an uncached rendering of the page will process all terms and trying to resolve all ObjectStorages. Even if there is *no* entry, this causes a SQL-request. As the normal has `description` and `media` as properties, there are too avoidable SQL-request per term.
For getting the terms cachable in a transient backend, lazyloading is *not* allowed due to serialization. Thus a SimpleTerm-object has been introduced without `description`, `media` and some simple properties which *is* cachable.